6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 /**
- * main - A program that prints the size of various computer
+ * main - computer size
  * Return: 0
  */
 int main(void)
 {
-printf("Size of a char: %zu byte(s)\n", (unsigned long)sizeof(char));
-printf("Size of an int: %zu byte(s)\n", (unsigned long)sizeof(int));
-printf("Size of a long int: %zu byte(s)\n", (unsigned long)sizeof(long int));
-printf("Size of a long long int: %zu byte(s)\n", (unsigned long)sizeof(long long int));
-printf("Size of a float: %zu byte(s)\n", (unsigned long)sizeof(float));
+printf("Size of a char: %zu byte(s)\n", sizeof(char));
+printf("Size of an int: %zu byte(s)\n", sizeof(int));
+printf("Size of a long int: %zu byte(s)\n", sizeof(long int));
+printf("Size of a long long int: %zu byte(s)\n", sizeof(long long int));
+printf("Size of a float: %zu byte(s)\n", sizeof(float));
 return (0);
 }
